% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_agency.R
\name{add_agency}
\alias{add_agency}
\title{Add a transit agency to the GTFS}
\usage{
add_agency(agency_name, agency_url, agency_timezone)
}
\arguments{
\item{agency_name}{Agency name}

\item{agency_url}{Agency URL link}

\item{agency_timezone}{Agency time zone}
}
\value{
Adds a new row to the \code{Agency} object corresponding to the new agency's information.
}
\description{
Add agency information (name, URL, time zone).
}
\examples{
add_agency("TaM Montpellier", "blabla.com", "Paris/Europe")
}
