% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_route.R
\name{add_route}
\alias{add_route}
\title{Add a route to the GTFS}
\usage{
add_route(name, type)
}
\arguments{
\item{name}{name of the route}

\item{type}{type of transportation used on a route:
\itemize{
\item \code{0}: Tram, Streetcar, Light rail. Any light rail or street level system within a metropolitan area
\item \code{1}: Subway, Metro. Any underground rail system within a metropolitan area
\item \code{2}: Rail. Used for intercity or long-distance travel
\item \code{3}: Bus. Used for short- and long-distance bus routes
\item \code{4}: Ferry. Used for short- and long-distance boat service
\item \code{5}: Cable tram. Used for street-level rail cars where the cable runs beneath the vehicle (e.g., cable car in San Francisco)
\item \code{6}: Aerial lift, suspended cable car (e.g., gondola lift, aerial tramway). Cable transport where cabins, cars, gondolas or open chairs are suspended by means of one or more cables
\item \code{7}: Funicular. Any rail system designed for steep inclines
\item \code{11}: Trolleybus. Electric buses that draw power from overhead wires using poles
\item \code{12}: Monorail. Railway in which the track consists of a single rail or a beam
}}
}
\value{
Add a new row to the \code{Routes} object corresponding to the new routes's information.
}
\description{
Add a new route
}
\examples{
add_route(name = "Montcalm - Lunaret",
          type = 0)
}
