% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stop_times.R
\name{add_stop_times}
\alias{add_stop_times}
\title{Add trip information to the GTFS}
\usage{
add_stop_times(trip_id_v, time, stop_id, stop_sequence)
}
\arguments{
\item{trip_id_v}{identification number of the trip concerned}

\item{time}{time "HH:MM:SS" at the stop for the specific trip}

\item{stop_id}{identification number of the stop concerned}

\item{stop_sequence}{stop order number for the trip concerned}
}
\value{
Add a new row to the \code{stop_times} object.
}
\description{
add trip information (time, stop, sequence) to stop_times
}
\examples{
add_stop_times(1, "05:00:00", 1, 1)
}
