% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trip.R
\name{add_trip}
\alias{add_trip}
\title{Add a trip to the GTFS}
\usage{
add_trip(route_id, service_id)
}
\arguments{
\item{route_id}{identification number of the route concerned}

\item{service_id}{identification number of the service concerned}
}
\value{
Add a new row to the \code{trips} object.
}
\description{
add a trip
}
\examples{
add_trip(route_id = 1,
          service_id = 1)
}
