% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trips.R
\name{add_trips}
\alias{add_trips}
\title{Add a trip to the GTFS}
\usage{
add_trips(route_id, service_id)
}
\arguments{
\item{route_id}{identification number of the route concerned}
}
\value{
Add a new row to the \code{trips} object.
}
\description{
add a trip
}
\examples{
add_trips(route_id = 1
          service_id = 1)
}
