% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_files.R
\name{gtfs_files}
\alias{gtfs_files}
\title{Create all files needed for GTFS}
\usage{
gtfs_files()
}
\value{
Create empty objets: \code{agency}, \code{stops}, \code{routes}, \code{trips}, \code{stop_times} and \code{calendar}
}
\description{
Create all files needed for GTFS
}
\examples{
gtfs_files()
}
