% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_save.R
\name{gtfs_save}
\alias{gtfs_save}
\title{Save the GTFS on disk}
\usage{
gtfs_save(filename)
}
\arguments{
\item{filename}{File name to create on disk}
}
\value{
Save six .txt files (agency, calendar, routes, stoptimes, stops, trips) on disk
}
\description{
Save the .txt files on disk
}
\examples{
\dontrun{
gtfs_save(filename = "C:/...")
}
}
