% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot a map}
\usage{
plot_map(route_id)
}
\arguments{
\item{route_id}{Identification of the route}
}
\value{
A leaflet map
}
\description{
Plot a map
}
\examples{
\dontrun{
plot_map(1)
}
}
