% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trips.R
\name{plot_trips}
\alias{plot_trips}
\title{Plot transport offer}
\usage{
plot_trips(route_id, size_point, size_line)
}
\arguments{
\item{route_id}{identification number of the route concerned}

\item{size_point}{numerical value of the point size}

\item{size_line}{numerical value of the line size}
}
\value{
Create a graph of the transport offer
}
\description{
Plot transport offer
}
\examples{
plot_trips(route_id = 1,
size_point = 1.4,
size_line = 0.8)
}
