% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_stop_times.R
\name{repeat_stop_times}
\alias{repeat_stop_times}
\title{Repeat a trip sequence for other timetables}
\usage{
repeat_stop_times(trip_id, max_time, breaks)
}
\arguments{
\item{trip_id}{identification number of the trip concerned}

\item{max_time}{The maximum time that must not be exceeded}

\item{breaks}{The number of minutes between two trips until the maximum time is reached}
}
\value{
Add new rows to the \code{stop_times} object.
}
\description{
Repeat a sequence of journeys for other times according to a specific frequency
}
\examples{
repeat_stop_times(trip_id = 1,
                 max_time = "20:00:00",
                 breaks = 10)
}
