% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_agency.R
\name{add_agency}
\alias{add_agency}
\title{Add a transit agency to the GTFS}
\usage{
add_agency(name, url, timezone)
}
\arguments{
\item{name}{Agency name}

\item{url}{Agency URL link}

\item{timezone}{Agency time zone}
}
\value{
Adds a new row to the \code{Agency} object corresponding to the new agency's information.
}
\description{
Add agency information (name, URL, time zone).
}
\examples{
add_agency("TaM Montpellier", "blabla.com", "Paris/Europe")
}
