% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_calendar.R
\name{add_calendar}
\alias{add_calendar}
\title{Add a calendar to the GTFS}
\usage{
add_calendar(
  service_id,
  monday,
  tuesday,
  wednesday,
  thursday,
  friday,
  saturday,
  sunday,
  start_date,
  end_date
)
}
\arguments{
\item{service_id}{Identification number of the service concerned}

\item{monday}{Indicates whether the service operates every Monday (0 if unavailable, 1 if available)}

\item{tuesday}{Indicates whether the service operates every Tuesday (0 if unavailable, 1 if available)}

\item{wednesday}{Indicates whether the service operates every Wednesday (0 if unavailable, 1 if available)}

\item{thursday}{Indicates whether the service operates every Thursday (0 if unavailable, 1 if available)}

\item{friday}{Indicates whether the service operates every Friday (0 if unavailable, 1 if available)}

\item{saturday}{Indicates whether the service operates every Saturday (0 if unavailable, 1 if available)}

\item{sunday}{Indicates whether the service operates every Sunday (0 if unavailable, 1 if available)}

\item{start_date}{Start service day for the service interval}

\item{end_date}{End service day for the service interval (this service day is included in the interval)}
}
\value{
Add a new row to the \code{Calendar} object corresponding to the new calendar's information.
}
\description{
Add calendar information (number of the service concerned, every day of the week, service start and end dates).
}
\examples{
add_calendar(1, 1, 1, 1, 1, 1, 1, 1, 20240511, 20241231)
}
