% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stops.R
\name{add_stops}
\alias{add_stops}
\title{Add a stop to the GTFS}
\usage{
add_stops(lat, lon, name)
}
\arguments{
\item{lat}{latitude of the location}

\item{lon}{longitude of the location}

\item{name}{name of the stop}
}
\value{
Add a new row to the \code{stops} object.
}
\description{
add a stop
}
\examples{
add_stops(lat = 43.59690,
         lon = 3.86357,
         name = "Parc Montcalm")
}
