% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stops.R
\name{add_stops}
\alias{add_stops}
\title{Add a stop to the GTFS}
\usage{
add_stops(stop_lat, stop_lon, stop_name)
}
\arguments{
\item{stop_lat}{latitude of the location}

\item{stop_lon}{longitude of the location}

\item{stop_name}{name of the stop}
}
\value{
Add a new row to the \code{stops} object.
}
\description{
add a stop
}
\examples{
add_stops(43.59608069974394, 3.8626826404232526, "Parc Montcalm")
}
