% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trips.R
\name{plot_trips}
\alias{plot_trips}
\title{Plot transport offer}
\usage{
plot_trips(service_id)
}
\arguments{
\item{service_id}{identification number of the service concerned}
}
\value{
Create a graph of the transport offer
}
\description{
Plot transport offer
}
\examples{
plot_trips(service_id = 1)
}
