% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_stop_times.R
\name{repeat_stop_times}
\alias{repeat_stop_times}
\title{Repeat a trip sequence for other timetables}
\usage{
repeat_stop_times(trip_id_n, horaire, temps)
}
\arguments{
\item{trip_id_n}{identification number of the trip concerned}

\item{horaire}{The maximum time that must not be exceeded}

\item{temps}{The number of minutes between two trips until the maximum time is reached}
}
\value{
Add new rows to the \code{stop_times} object.
}
\description{
Repeat a sequence of journeys for other times according to a specific frequency
}
\examples{
repeat_stop_times(1, "23:00:00", 7)
}
